pyproject.toml
[project]
name = "aws-helper-bot"
version = "0.1.0"
description = ""
authors = [
    {name = "irayspace",email = "irayspacii@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "discord-py (>=2.5.2,<3.0.0)",
    "boto3 (>=1.38.3,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
